# notes/stylenest-openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "product-controller"
                ],
                "operationId": "getProductById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseProductResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "product-controller"
                ],
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseProductResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "product-controller"
                ],
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/products/{id}": {
            "get": {
                "tags": [
                    "admin-product-controller"
                ],
                "operationId": "getProductById_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseProductResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "admin-product-controller"
                ],
                "operationId": "updateProduct_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseProductResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "admin-product-controller"
                ],
                "operationId": "deleteProduct_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/products": {
            "get": {
                "tags": [
                    "product-controller"
                ],
                "operationId": "getAllProducts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseListProductResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "product-controller"
                ],
                "operationId": "createProduct",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseProductResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/products": {
            "get": {
                "tags": [
                    "admin-product-controller"
                ],
                "operationId": "getAllProducts_1",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseListProductResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "admin-product-controller"
                ],
                "operationId": "createProduct_1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseProductResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProductRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 1
                    },
                    "shortDescription": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number",
                        "minimum": 0.0
                    },
                    "discountPrice": {
                        "type": "number"
                    },
                    "fabric": {
                        "type": "string"
                    },
                    "careInstructions": {
                        "type": "string"
                    },
                    "featured": {
                        "type": "boolean"
                    },
                    "active": {
                        "type": "boolean"
                    },
                    "categoryId": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "required": [
                    "categoryId",
                    "name",
                    "price"
                ]
            },
            "ApiResponseProductResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/ProductResponse"
                    }
                }
            },
            "ProductResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "shortDescription": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "discountPrice": {
                        "type": "number"
                    },
                    "fabric": {
                        "type": "string"
                    },
                    "careInstructions": {
                        "type": "string"
                    },
                    "featured": {
                        "type": "boolean"
                    },
                    "active": {
                        "type": "boolean"
                    },
                    "thumbnailUrl": {
                        "type": "string",
                        "description": "URL of the first image (lowest displayOrder) belonging to the product's first variant that has images. Null if the product has no variant images yet."
                    },
                    "availableColors": {
                        "type": "array",
                        "description": "Distinct colors across this product's variants. Empty (not null) if the product has no variants.",
                        "items": {
                            "type": "string"
                        }
                    },
                    "categoryName": {
                        "type": "string"
                    },
                    "gender": {
                        "type": "string",
                        "description": "The customer segment this product belongs to, inherited from its category.",
                        "enum": [
                            "MEN",
                            "WOMEN"
                        ]
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "ApiResponseVoid": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {}
                }
            },
...
            "ApiResponseListProductResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductResponse"
                        }
                    }
                }
            },
...
        }
    }
}